+++ src/ConfigElementType/CommentConfigElementType.php
@@ -18,7 +18,7 @@ use HeimrichHannot\ReaderBundle\Backend\ReaderConfigElement;
 use HeimrichHannot\ReaderBundle\Item\ItemInterface;
 use HeimrichHannot\ReaderBundle\Model\ReaderConfigElementModel;
 
-class CommentConfigElementType implements ConfigElementType
+class CommentConfigElementType implements ReaderConfigElementTypeInterface
 {
     /**
      * @var ContaoFrameworkInterface
@@ -52,6 +52,36 @@ class CommentConfigElementType implements ConfigElementType
         }
 
         $item->setFormattedValue('comments', $template->parse());
+    }
+
+    /**
+     * Return the reader config element type alias.
+     *
+     * @return string
+     */
+    public static function getType(): string
+    {
+        return 'comment';
+    }
+
+    /**
+     * Return the reader config element type palette.
+     *
+     * @return string
+     */
+    public function getPalette(): string
+    {
+        return '{config_legend},commentTemplate,commentCustomTemplate,commentNotify,commentSortOrder,commentPerPage,commentModerate,commentBbcode,commentRequireLogin,commentDisableCaptcha,commentOverridePalette,commentHideFields;';
+    }
+
+    /**
+     * Update the item data.
+     *
+     * @param ReaderConfigElementData $configElementData
+     */
+    public function addToListItemData(ReaderConfigElementData $configElementData): void
+    {
+        $this->addToItemData($configElementData->getItem(), $configElementData->getReaderConfigElement());
     }
 
     /**