+++ src/DependencyInjection/Compiler/ReaderCompilerPass.php
@@ -1,0 +1,38 @@
+<?php
+
+/*
+ * Copyright (c) 2019 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+namespace HeimrichHannot\ReaderBundle\DependencyInjection\Compiler;
+
+use HeimrichHannot\ReaderBundle\Registry\ReaderConfigElementRegistry;
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Reference;
+
+class ReaderCompilerPass implements CompilerPassInterface
+{
+    /**
+     * You can modify the container here before it is dumped to PHP code.
+     */
+    public function process(ContainerBuilder $container)
+    {
+        // always first check if the primary service is defined
+        if (!$container->has(ReaderConfigElementRegistry::class)) {
+            return;
+        }
+
+        $definition = $container->findDefinition(ReaderConfigElementRegistry::class);
+
+        // find all service IDs with the app.mail_transport tag
+        $taggedServices = $container->findTaggedServiceIds('huh.reader.config_element_type');
+
+        foreach ($taggedServices as $id => $tags) {
+            // add the transport service to the TransportChain service
+            $definition->addMethodCall('addReaderConfigElementType', [new Reference($id)]);
+        }
+    }
+}