+++ src/DependencyInjection/ReaderExtension.php
@@ -33,6 +33,7 @@ class ReaderExtension extends Extension
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
 
         $loader->load('services.yml');
+        $loader->load('datacontainers.yml');
         $loader->load('listener.yml');
     }
 }